requirements.txt
# Core
python-dotenv

# UI
streamlit

# LangChain core
langchain
langchain-core
langchain-community
langchain-text-splitters

# Groq LLM
langchain-groq

# Vector DB
faiss-cpu

# Docs
pypdf
